Check container naming constraints
The plugin uses the image name as container name. However,
there are some restrictions on container names which does
not apply to KIWI image names. Thus we have to check if
the name in the image description can be used as container
name and if not come up with an error message and point
to the new --container-name option to provide an alternative
name