Move stash home to /var/tmp/kiwi-stash
This is the better path as the stash container is locally
only kept to add more layers to it. The location in /var/tmp
indicates it could also be deleted which is true since every
stash gets registered to the local podman registry. In
addition the location change allows the --list command to
operate without root permissions